Project structure refactoring continue
