+++ src/components/Menu/MenuOptions.tsx
@@ -4,7 +4,7 @@ import AboutMenu from '@components/Menu/AboutMenu';
 import ImportExportChat from '@components/Chat/ImportExportChat';
 import SettingsMenu from '@components/Menu/SettingsMenu';
 import CollapseOptions from './CollapseOptions';
-import TokensCostMenu from './TokensCostMenu'
+import TokensCostModal from './TokensCostModal'
 import AuthenticatedUserLogout from './AuthenticatedUserLogout';
 
 const googleClientId = import.meta.env.VITE_GOOGLE_CLIENT_ID || undefined;
@@ -20,7 +20,7 @@ const MenuOptions = () => {
           hideMenuOptions ? 'max-h-0' : 'max-h-full'
         } overflow-hidden transition-all`}
       >
-        {countTotalTokens && <TokensCostMenu />}
+        {countTotalTokens && <TokensCostModal />}
         <SettingsMenu />
         <AboutMenu />
         <AuthenticatedUserLogout />
