--- src/components/Menu/MenuOptions/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MenuOptions';
